+++ semcore/utils/src/use/usePreventScroll.tsx
@@ -24,27 +24,40 @@ export function getScrollbarWidth() {
   return widthNoScroll - widthWithScroll;
 }
 
+function getIntValueFromCss(value) {
+  return !Number.isNaN(Number(value)) ? Number(value) : Number.parseInt(value, 10);
+}
+
 export default function usePreventScroll(visible = true) {
   const scrollbarWidth = useMemo(getScrollbarWidth, [getScrollbarWidth]);
   const paddingRightRef = useRef<string>(null);
   const overflowRef = useRef<string>(null);
+  const boxSizingRef = useRef<string>(null);
+
   useEffect(() => {
     if (!canUseDOM() || !visible) return;
 
-    const { overflow, paddingRight } = window.getComputedStyle(document.body);
+    const { overflow, paddingRight, boxSizing } = window.getComputedStyle(document.body);
     paddingRightRef.current = paddingRight;
     overflowRef.current = overflow;
+    boxSizingRef.current = boxSizing;
 
-    const intPaddingRight = !Number.isNaN(Number(paddingRight))
-      ? Number(paddingRight)
-      : Number.parseInt(paddingRight, 10);
+    const intPaddingRight = getIntValueFromCss(paddingRight);
+    let intPaddingRightFromStyle = getIntValueFromCss(document.body.style.paddingRight);
+    // Detected own style for window inside window
+    if (intPaddingRightFromStyle !== scrollbarWidth) {
+      intPaddingRightFromStyle = 0;
+    }
 
     document.body.style.overflow = 'hidden';
-    document.body.style.paddingRight = scrollbarWidth + intPaddingRight + 'px';
+    document.body.style.paddingRight =
+      scrollbarWidth + (intPaddingRight - intPaddingRightFromStyle) + 'px';
+    document.body.style.boxSizing = 'border-box';
 
     return () => {
       document.body.style.overflow = overflowRef.current;
       document.body.style.paddingRight = paddingRightRef.current;
+      document.body.style.boxSizing = boxSizingRef.current;
     };
   }, [visible]);
 }